test: Rename from tests
